+++ packages/nimble-components/src/text-area/tests/text-area.stories.ts
@@ -12,6 +12,8 @@ interface TextAreaArgs {
     value: string;
     readonly: boolean;
     disabled: boolean;
+    errorVisible: boolean;
+    errorText: string;
     spellcheck: boolean;
     resize: TextAreaResize;
     rows: number;
@@ -19,14 +21,19 @@ interface TextAreaArgs {
     maxlength: number;
 }
 
+const description = `
+A multi-line text input control. The text area is often used in a form to collect user inputs like comments or reviews.
+
+If you configure your text area to be resizable (with the \`resize\` attribute) in a certain dimension, do not set an explicit size for that dimension (via \`height\` and/or \`width\` \`style\` properties), or you may experience unexpected resize behavior. If you want to set the initial size of a resizable text area, use the \`rows\` and/or \`cols\` attribute(s).
+`;
+
 const metadata: Meta<TextAreaArgs> = {
     title: 'Text Area',
     tags: ['autodocs'],
     parameters: {
         docs: {
             description: {
-                component:
-                    'A multi-line text input control. The text area is often used in a form to collect user inputs like comments or reviews.'
+                component: description
             }
         },
         actions: {
@@ -40,6 +47,8 @@ const metadata: Meta<TextAreaArgs> = {
             :value="${x => x.value}"
             ?readonly="${x => x.readonly}"
             ?disabled="${x => x.disabled}"
+            ?error-visible="${x => x.errorVisible}"
+            error-text="${x => x.errorText}"
             spellcheck="${x => x.spellcheck}"
             resize="${x => x.resize}"
             rows="${x => x.rows}"
@@ -58,7 +67,8 @@ const metadata: Meta<TextAreaArgs> = {
             }
         },
         resize: {
-            description: 'Direction(s) the text area is sizeable by the user',
+            description:
+                'Direction(s) the text area is sizeable by the user. Setting a fixed `height` and `width` on the text area is not supported while it is sizeable. You may instead use `rows` and `cols` to set an initial size.',
             options: Object.values(TextAreaResize),
             control: { type: 'select' },
             table: {
@@ -78,6 +88,14 @@ const metadata: Meta<TextAreaArgs> = {
         maxlength: {
             description:
                 'Maximum number of characters that may be entered by the user'
+        },
+        errorVisible: {
+            description:
+                'Whether the text area should be styled to indicate that it is in an invalid state'
+        },
+        errorText: {
+            description:
+                'A message to be displayed when the text area is in the invalid state explaining why the value is invalid'
         }
     },
     args: {
@@ -87,6 +105,8 @@ const metadata: Meta<TextAreaArgs> = {
         value: loremIpsum,
         readonly: false,
         disabled: false,
+        errorVisible: false,
+        errorText: 'Value is invalid',
         spellcheck: false,
         resize: TextAreaResize.both,
         rows: 3,